[FD][APB-739] Use "fire and forget" approach to auditing
...as per the example in the https://github.com/hmrc/play-auditing README.

So as not to impact the performance of our service if things are slow/going wrong in the audit pipeline.

If an error occurs whilst play-auditing is sending the event to datastream it will be logged and the audit-sweep service will scan the logs and retry the audit event at a later date.